SEACAS: Remove HDF5 requirement from exodus
After further Direction from Greg Sjaardema I am removing the HDF5 requirement
for exodus. According to Greg the actual required netcdf is just 4.1.3. However,
a netcdf requiring hdf5 will be needed in the future. For now though I am
removing the hard requirement that I added into SEACASExodus previously to allow
the use of a netcdf that is not built with HDF5 support.