Adding getList() helper function for getting the array of integers supplied for an integer range list parameter.
const Array<int> *paramValue = params.getPtr<Array<int> >("number_list");
ArrayView<const int> paramView = getList(paramValue);

size_t numValues = paramView.size();
int firstValue = paramView[0];

paramValue is encoded list, which contains a flag indicating that the
value is "all" or "none" or a list of integers.

getList returns a view of the array of integers, without the flag.

If the user indicated "all" or "none", then the size of the list is 0.