Reproduced and fixed a nasty defect related to inconsistent extra repos files and input list
This commit adds a unit test that exposes a nasty defect we just started
seeing in VERA where we added a new extra repo that had the PACKSTAT set to
NOPACKAGES in the cmake/ExtraRepositoriesList.cmake file and then we started
getting extra respos specified using -DVERA_EXTRA_REPOSITORIES on the CMake
command line ignored where it complained that an extra repo was being ignored
that had PACKSTAT set to NOPACKAGES.  These should be unrelated but they were
not.

It turned out that there was bad logic in the TriBITS cmake code where the
varaible ${PROJECT_NAME}_EXTRA_REPOSITORIES_PACKSTATS read from the extra
repos file was being preserved and used even though the list of extra repos
being used could be totally different when passed in through
VERA_EXTRA_REPOSITORIES.  The unit test I added exposed this behavior so that
I could be sure that I fixed it.

To fix this, I added logic to see if the list of extra repos passed in
VERA_EXTRA_REPOSITORIES differed from what was read from the extra repos file,
and if it is, then the data read from the extra repos file is wiped out.  I
then had to fix some logic that was reading
${PROJECT_NAME}_EXTRA_REPOSITORIES_PACKSTATS to only read it if it was
non-empty (which would only now be the case if the list of extra repos can
from the extra repos file).

After this, the behavior of using VERA_EXTRA_REPOSITORIES with VERA is working
just fine and is now well protected due to the new unit test.  What a pain.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework, Panzer
Disabled Packages: PyTrilinos,Optika,Sundance,Stokhos
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=200,notpassed=0 (119.00 min)