SEACAS: Fix script installation source paths
The previous scripts had to be configured and as such were in the binary tree.
With the change to the scripts not being configured they exist only in the
source tree. I had just cut and pasted the variables that were used to find the
decomp script for the others which was using the current binary dir instead of
the proper current source dir. They now correctly point to the source dir for
the scripts to install