Break dependency on MueLu in Xpetra
Xpetra had a couple dependencies on MueLu files. A couple examples needed to be
commented out as they required MueLu sources to build. A MueLu header was being
included in an Xpetra file. It appears that it wasn't actually used though since
commenting it out allowed the file to compile.