Excluding Xpetra and MueLu from CASL builds of Trilinos
When Xpetra and MueLu got moved from preCoprightTrilinos into Trilinos proper
as PS packages, the optional dependency in Drekar on MueLu caused them to be
built as part of CASL VERA builds.  This immediately resulted in build errors
in Xpetra and MueLu.  While the Xpetra errors seems to have gone away, the
MueLu errors persist.  While MueLu may be important for CASL in the future, it
is not being used right now.  After asking Roger Pawlowski about this, we
agreed to disable MueLu (and Xpetra) in CASL-related builds for now.  This
means that MueLu and Xpetra will not be built or tested in Trilinos-based or
VERA-based TriBITS builds run by CASL.

Build/Test Cases Summary
Enabled Packages: Drekar
Disabled Packages: PyTrilinos,Optika,Sundance,Stokhos
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=23,notpassed=0 (103.89 min)