Panzer: Now handles null operators in blocked scatter
This is for cases where you now a priori that you don't need
to scatter a particular block pair. You can specify which blocks
are excluded by giving a list of pairs of integers. For instance
if you want to exclude off diagonal coupling in one direction
for a two by two operator you would write
          <Parameter name="Excluded Blocks" type="string" value="0,1"/>
if you want to exclude both off diagonal couplings, write
          <Parameter name="Excluded Blocks" type="string" value="0,1;1,0"/>

This commit also includes an additional convenience function in
the string utilities.