Changed disable of sacado definition of c++11 math functions.
This commit addresses bug #4858.  Sacado defines some math functions
that don't normally appear in cmath, but are now part of c++11.  For a
while sacado has had configure checks for these and disables them if
they are defined.  However this only works if sacado is compiled (say in
an application code) with the same cxxflags used to configure trilinos.
This commit adds additional preprocessor guards to disable defining
these functions at compile time, regardless of how trilinos was
configured.  They are disable if sacado is compiled with "g++
-std=c++0x", or if the user puts -D USER_DISABLE_SACADO_TR1_CMATH
in their cxxflags.