Make it an error to enable a package with a req. dep. on something disabled.
Before it was only a warning to explicitly enable a package that had a required
dependency on something (a package or tpl) that was disabled. The package would
announce that it was being disabled, but the configure would continue. This is
particularly an issue when two conflicting options are given and no sane
configure can realisitically be done, for instance turning on Panzer in a serial
build(which now explicitly disables MPI.) The configure will now stop at the
point where the package is being disabled even though it was explicitly set to
ON so it should be clear to users what is wrong with the configure.

Note, the error message had to be modified slightly beyond making it cause cmake
to stop due to an odd change in cmake's behavior when printing error messages vs
standard or status messages. Apparently error messages are formated to be
indented and line wrapped at about 76 chars where as standard messages are
printed verbatim. To get around this I had to indent the error message myself,
which tells cmake not to line wrap, but it still indents. The indention wasn't a
large issue, but the line wrapping was as it broke up a string that was being
used for unit tests to compare against.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
0) MPI_DEBUG => passed: passed=102,notpassed=0 (12.60 min)
1) SERIAL_RELEASE => passed: passed=102,notpassed=0 (3.56 min)