IOSS/FRIO: Add append option for results, restart, history, and heartbeat dbs (only if no topology change)
The command line to activate this is:
   EXISTS = {overwrite|append|abort}

If 'abort' is specified, the code will stop with an error message if
a file with the same name exists.

If 'overwrite' is specified, an existing file will be overwritten.
This is the default for all file types except that HEARTBEAT and HISTORY
default to 'append'

If 'append' is specified AND the run is a restart run, then new data
will be appended to the end of an existing file.  If it is not a restarted
run, then an existing file will be overwritten.  Note that there is currently
no check that the mesh topology of an existing database matches the current
mesh topology, so it is currently up to the user to ensure that appending
to an existing database is valid.

(From Sierra)