Teuchos: Bypassed generic reduceAll for Ordinal=int and Packet=int,double.
This is a temporary fix to address Bug 5240.  It may also help with Bug 5682.

https://software.sandia.gov/bugzilla/show_bug.cgi?id=5240
https://software.sandia.gov/bugzilla/show_bug.cgi?id=5682

I added specializations of the "comm helper" nonmember function reduceAll for
Ordinal=int (which is always the case for current MPI implementations) and
Packet=int,double.  The specializations use the correct MPI_Datatype: MPI_INT
for int, and MPI_DOUBLE for double.  This should avoid the incorrect splitting
of large reduction arrays observed in Bug 5240.

The specializations attempt to cast first directly to MpiComm (if building with
MPI), then directly to SerialComm.  If neither works, they fall back to the
current generic implementation that relies on serialization.

If this specialization approach works, I will add specializations for other
popular Packet types.  This should buy us time to come up with a more general
solution.

Build/Test Cases Summary
Enabled Packages: Teuchos
0) MPI_DEBUG => FAILED: passed=95,notpassed=2 => Not ready to push! (0.47 min)
1) SERIAL_RELEASE => passed: passed=95,notpassed=0 (0.33 min)
WARNING: Forced the push!