Refactoring of Teuchos to use Subpackages (Trilinos bug 5738)
To see the motivation and details for this, see Trilinos bug 5738.

This commit should maintain 100% backward compatibility except for going from
one lib 'teuchos' to multiple libs.  Such is the right of every Trilinos
package however.

The raw intermediate commits made to do this refactoring as shown in the
teuchos-subpackages-20121025-final branch in the git repo:

   software.sandia.gov:/home/rabartl/git/Trilinos

The commits were squashed here to create one clean commit to make easy to back
out and there is no reason to leave the messy history long term.

Here is a listing of the individual commit messages that you will find on that
branch:

Initial refactoring for TeuchosCore subpackage (SQUASH AGAINST THIS).

More incremental subpackage work (SQUASH THIS).

Fixed paths in test output grep for move to core subpackage (SQUASH THIS).

Adding TeuchosParameterList subpackage (SQUASH THIS).

Mostly got TeuchosComm subpackage to work (SQUASH THIS).

Fixed the dependency between TimeMonitor and CommandLineProcessor by inverting the dependency using C++ magic (SQUASH THIS).

Adding CWrapperSupport to TeuchosCore subpakage.

Forgot to move Condition tests to PL subpackage; Moved CTeuchosMonitor to Comm subpackage.

Created TeuchosNumerics subpackage (SQUASH THIS)

Moving Hashtable code to core since it is used in Trilinos (SQUASH THIS)

Adding Polynomial code to Numerics subpackage because it it is not big, it is tested and it related to numerics.

Moved rest of code to Remainder subpackage.

Removing single-package files and dirs (SQUASH THIS).

Basic update of Doxygen documentation (SQUASH THIS).