Teuchos: Several changes to Comm and related classes: - Improved documentation of Comm, CommRequest, and CommStatus. - CommRequest now includes a virtual destructor.  Subclasses' implementations   of the destructor are supposed to cancel the pending communication request.   (In MPI, this behaves as a local operation.  See the implementation in   MpiCommRequest.) - The CommRequest interface now includes a wait() method that returns a   CommStatus<OrdinalType> (just as MPI_Wait returns an MPI_Status). - CommRequest (and its subclass MpiCommRequest) now takes an OrdinalType   template parameter, corresponding to the template parameter of Comm.  This   lets CommRequest::wait() return a CommStatus with the correct template   parameter.
The last point is an incompatible interface change that may affect direct users
of Teuchos::Comm.  I expect the primary customers of Teuchos::Comm to be Tpetra
and Zoltan2, with perhaps some use by MueLu.  I have built Zoltan2 and MueLu,
and did not need to make changes to them.  (In fact, I have built all
downstream packages of Teuchos, with no problems.)  Tpetra needed only a few
minor changes to the implementation of Distributor, CrsGraph, and CrsMatrix.
Tpetra uses but does not expose CommRequest, so Tpetra users will not be
affected.  I will check in the Tpetra changes in my next commit.