Amesos: Protect the CSparse code when CSparse isn't available
Sierra was running into an issue with Amesos_CSparse because only parts of the
Amesos_CSparse.h were protected if CSparse wasn't enabled. The Trilinos build
system did not build these files if CSparse wasn't available so we never saw
this on our end. I moved the #ifdef protections to cover the entire files
Amesos_CSparse.h and cpp to make it so Sierra won't have to make an exception
for their build system for these files.