Bug #5349
Adding support for "long long int" in the Teuchos command-line processor tool.
This was aided by Mark's addition of string to long int and string to long long int support
in the ValueTypeConversionTraits template class.  The example has been extended to check
that "long long int" can be used with the CommandLineProcessor.  Like much of the
long long int support in Teuchos, this will only be enabled when Trilinos is configured
with the cmake option Teuchos_ENABLE_LONG_LONG_INT:BOOL=ON.