Tpetra: CrsMatrix: Implemented relaxed-requirement gaussSeidelCopy.
The gaussSeidelCopy method is like gaussSeidel, except that the
input/output multivector X need not be a domain Map view of a column
Map multivector.  X need only be a domain Map multivector.  As a
result, the method must copy into a temporary multivector for the
Gauss-Seidel kernel.  This gives us a gradual migration path for
Ifpack2, which doesn't currently require that the input multivector
of its Gauss-Seidel be a view of a column Map multivector.