Belos: Fixed MV test to allow for nondeterminism in norm results.
The MultiVecTraits test was assuming that computing the 2-norm of
the same vector twice would always result in exactly the same value.
This is not necessarily true (due to floating-point rounding error
and the MPI standard allowing the MPI implementation to use different
reduction trees even for reductions with the same processes) even for
an MPI-only implementation, but it was showing up especially in Belos'
Tpetra tests (which use both MPI and threads).

There are two ways to fix this problem correctly.  First, if the
vectors really are supposed to be identical (as they are for many of
the tests), then the infinity norm (max of the absolute values of the
difference between the two vectors) is the right metric.  Second, if
the two-norm really is the right metric, then one should use error
analysis to figure out how big the difference should be.

Instead of implementing the above solutions, I "fixed" the problem
by introducing a small arbitrary (100*eps) error tolerance and
replacing all "norms2[i] != norms[i]" type comparisons with absolute
magnitude comparisons to the error tolerance.  It's a hack and we
should contemplate revisiting this test, but it's good enough for
now to catch any serious issues with a MultiVecTraits implementation
(which is the point of the unit test anyway).