Bug # 5756
Adding a MultiVecTraitsExt class to augment the current MultiVecTraits class
with a GetGlobalLength() call which returns a ptrdiff_t.  This will address
Anasazi's interface to Epetra64 in a backward compatible way.  Default definitions
are given for this method in the traits and inheritance multivector interfaces
so that those who have written their own interfaces to Anasazi's multivector classes
or written solvers on top of these classes, won't be affected by these changes.
The current Anasazi solvers have been modified to use the GetGlobalLength() method
for performing sanity (or compatibility) checks.  The only solver not modified was
the RTR solvers, which I will leave for Chris Baker to change at his leisure.

A test has been added, provided by Karen Devine, that uses Epetra64 functionality
to generate a matrix.