Panzer: Adding WorksetDescriptor and sideset volume workset construction
This is in support of higher order surface responses. Additional small
changes improving the interface to the stk mesh objects were made.

Adding a workset descriptor (name pending) abstracts away the need to
pass around the block id or sideset and allows more flexibility in specifying
the elements required in the workset.