Ifpack2: Chebyshev: Relaxed requirements on matrix and diagonal.
I relaxed the requirement that the row Map and the range Map must be
the same (in the sense of Tpetra::Map::isSameAs).  Now they may differ.

I also now let the user-supplied vector of diagonal entries of the
matrix have any Map.  The routine will do increasingly costly tests
to avoid building a new Export object.  In the common case where
range Map and row Map are identical (pointers to the same Map) and
the input vector's Map is a pointer to that Map, then the tests will
only do an RCP pointer comparison.

Finally, compute() now computes the vector of diagonal entries (if
the user didn't give us the vector) on every call, not just on the
first call to compute().  This allows changes to the diagonal entries
of the matrix between calls to compute().  Users should always call
compute() before calling apply() if the matrix has changed in any way.