Ifpack2: Added test for convergence of Chebyshev.
The implementation of Chebyshev iteration in Ifpack2 was different
from the one in "Templates for the Solution of Linear Systems...",
so I implemented the version from the latter (in the new unit test
file), and wrote a comparison.  Both versions do converge, though
differently.  The differences could relate to each version's
heuristic for adjusting the input spectral bounds for better
convergence, but I'm not sure.  However, the test does build and
print useful output when ctest is run with the --verbose option.