Ifpack2: Chebyshev: Convergence test compares vs. textbook algorithm.
The Chebyshev convergence test now compares Ifpack2's Chebyshev
against the following implementations:

1. A textbook version of the algorithm
2. A clean reimplementation of Ifpack2's algorithm
3. A textbook implementation of CG

All three do the same left diagonal scaling that Ifpack2 does.
"Textbook" implementations come from "Templates for the Solution
of Linear Systems," 2nd edition.  We include CG just to give us a
rough measure of how fast the methods "should" converge.

The test exercises all four algorithms with a 1-D Poisson equation.
We know the eigenvalues of the matrix exactly, so we can give
perfect eigenvalue bounds.  We also experiment with changing the
max eigenvalue bound so that it's either an overestimate or an
underestimate.

I'm a bit frustrated because #2 and Ifpack2's Chebyshev _almost_
agree, but not quite.  The textbook implementation converges faster
if the eigenvalue bounds are good, but both Ifpack2's Chebyshev and
its imitation are much less sensitive to an incorrect upper bound on
the eigenvalues.  This gives me confidence that Ifpack2's version is
correct.

I haven't yet done the analysis to understand why the different
implementations produce different results.  However, I feel better
now about Ifpack2's version.  I would still prefer to use the textbook
version of the algorithm, though, just to remove the puzzle of whether
what you see on the screen is actually what it claims to be.