Piro: Allowing users to register custom steppers with RythmosSolver
This is a simple implementation of the dynamic registration interface.
This also requires the inclusion of a new RythmosStepperFactory, an
abstract interface that is used by the RythmosSolver to satisfy
the Open/Closed principle.

Also added a new initialize method and default constructor that
makes using the new factory registration method possible.

Also added a unit test that excercises this new capability.