Ifpack2: Chebyshev: Replaced insides with new implementation.
Now the "hand-rolled new implementation of Chebyshev" to which I
referred in previous commits is the implementation of Chebyshev
iteration used by Ifpack2::Chebyshev.  Convergence has improved
when the user did not provide an estimate of the max eigenvalue,
and otherwise has not changed (as predicted).

I still need to update the Ifpack2::Chebyshev documentation to
reflect the new features, like automatic eigenanalysis if you
don't provide an estimate of the max eigenvalue.