Teuchos: Added TPL Eigen, a templated C++ library for matrix math.
Eigen code is currently optionally used by SerialDenseSolver, the new SerialQRDenseSolver,
and the new SerialBandDenseSolver (the Eigen code for SerialBandDenseSolver has not
been committed yet) to allow these classes to use scalar types other than the MKL supported
ones. The dense matrix (and vector) classes do not contain any Eigen code, but I wrote
experimental Eigen versions of these and some parts could be useful for efficiency reasons,
but this is not necessary for functionality now.