KokkosArray:  Added KokkosArray_ENABLE_DeviceCuda as a Tribits option.
This also creates a CPP macro called HAVE_KOKKOSARRAY_CUDA to
enable/disable code based on whether Cuda is enabled.  The default for
KokkosArray_ENABLE_DeviceCuda is now equal to TPL_ENABLE_CUDA, so that
Cuda support is enabled by default in KokkosArray if Cuda is enabled
globally.

This required creating a KokkosArray_config.h.in for cmake to
process.