Panzer: Resolving issue with unstructured ordering of worksets in initial condition
The initial condition construction was failing for a conjugate
heat transfer problem. The reason was that the initial condition
evaluation routine was dependent on the ordering being consistent
between the field managers and the worksets. In a recent change
to the workset container and general workset construction this assumption
was somehow broken. However, this assumption was fragile to say
the least, now the initial condition field managers are constructed
in such a way that they are directly associated with an element block
through a map. This is not really that expensive since evaluation only
occurs once and there is never a lookup, so much as an iteration.

Also removed the dangerous getVolumeWorksets() method from the workset
container that returned all the volume worksets in a somehow unspecified
order. Now you can only access the worksets by element block.