Piro: Enabled specifying a single reponse index
Added another overload of PerformSolve which allows to select the index
of the single response to be computed instead of defaulting to the first
or having to use the most general function.

For the sake of consistency and completeness, the most general overload
uses the given response index when "Index" is specified as the response
selection method.

Made the option available in PerformAnalysis.