Panzer: Fixing CHT bug in global unknown numbering
There was a bug that caused heterogeneous multiphysics
assembly to fail. RPP pointed this out for a conjugate
heat transfer problem, that worked for some processor
sizes but failed (rather spectacularly) for others.
The issue comes down to a case where an unknown was
assigned to a processor for ownership, but none of its
owned elements needed that GID. This only happens in
heterogeneous multi-physics assembly and would not have
been a problem for "single-physics".

Regardless this appears to fix the problem.