Tpetra: Fixed build warnings due to newly deprecated function.
In my last commit, I deprecated the old constructor of
RowMatrixTransposer, and added a new, safer constructor.
In this commit, I replaced Tpetra's calls to the old constructor
with calls to the new constructor.

This will fix build warnings in Tpetra, but not necessarily in other
packages that use RowMatrixTransposer.  However, RowMatrixTransposer
was written in summer 2011 as part of the implementation of Tpetra's
sparse matrix-matrix multiply, so it's doubtful that other packages
use it.