Panzer: Fixing incorrect construction of the Epetra_CrsGraph
This was an issue where a particular column map was supposed to be
prescribed, however, it was done incorrectly. A small change to
the method calling the constructor fixes this problem. This hadn't
been an issue before, it was primarily an issue for the heterogeneous
multi-physics capability (inparticular the CHT problem).