Kokkos: AltSparseOps: Added gaussSeidel() from DefaultHostSparseOps.
I did this so that we can use AltSparseOps as the default sparse
kernel set for Node=SerialNode, instead of DefaultHostSparseOps.
(I haven't done this yet, but I plan to do it soon.)  AltSparseOps
does not rely on the C++ compiler to inline into the loop body of
sparse matrix-vector multiply.  Intel 13 and modern gcc are good
at inlining, but Intel < 13 and gcc 4.4 are apparently not.

This is a hack for the ASC L2 Milestone.  We're using SerialNode
for the Milestone.  I want those kernels to go as fast as possible,
given that they don't use threads at all.