Ifpack2: Details::Chebyshev: Fixed eigenvalue bounds computation.
Please refer to the following commit:

a853f3620ffad581f60380710355193cdb3485e9

In the commit message, what Jonathan calls the "setup()" method is
actually the compute() method.  His changes introduced a case where
it was possible for the eigenvalue bounds never to get computed: if
compute() had never been called before, and the first call set the
optional argument assumeMatrixUnchanged=true.  I fixed this, cleaned
up the logic, added more and better error checking, and improved the
public and private documentation.

In a future commit, I will remove the optional assumeMatrixUnchanged
argument to compute(), and make it an optional Boolean parameter in
the ParameterList argument of setParameters(): "chebyshev: assume
matrix does not change".