Changing description for Details::Chebyshev
It's useful to have in the one-line description the parameters used in the apply, in addition to the class name.
While this information is in the describe method, there's a great deal of other information that is noise when
Chebyshev is used inside other methods.  So I'm adding the eigenvalue estimates and degree to the description string.