Xpetra: Fixed TpetraMap constructors to get the right default Node.
Kokkos::DefaultNode::getDefaultNode() only gets an instance of the
correct Node type if the Node template parameter is the same as
Kokkos::DefaultNode::DefaultNodeType.  Otherwise, using it will
cause a syntax error.

The Kokkos::Details::getNode() template function is templated on
the Node type, so it always gets an instance of the right type.

I've fixed the constructors (and nonmember constructors) of
Xpetra::TpetraMap to use Kokkos::Details::getNode().