Fixing projection of coordinates.
Coordinates were being projected with R.  However, because rows of R are normalized, the scaling of the
projected coordinates was off.  This in turn cause the graph partitioner to redistribute the matrix
poorly, which in turn broke the aggregation.

Big thanks to Chris Siefert to helping to track this down.