STK_IO: Allow application to decide how to output nodal fields defined on parts
By default, a new nodeset will be created with the name part_name+"_nodes" and
the parts nodal fields will be output on that nodeset

If MeshData::use_nodeset_for_part_nodes_fields(false) is called,
the the part nodal fields will be output on all nodes of the model
and zero-filled where they aren't defined.  If two parts have the
same field defined on their nodes, then the field will be output
only once.

The nodeset will still be created so that the user can filter the nodal
variable by nodeset nodes in the visualization if needed.

The current default setting is that the fields will be output as
nodeset fields.

(From Sierra)